changed type

+++ Sources/BeaconChain.swift/BLS.swift
@@ -2,11 +2,11 @@ import Foundation
 
 class BLS {
 
-    static func verify(pubkey: Data, message: Data, signature: Data, domain: UInt64) -> Bool {
+    static func verify(pubkey: Data, message: Data, signature: Data, domain: Int) -> Bool {
         return true // @todo
     }
 
-    static func verify(pubkeys: [Data], messages: [Data], signatures: [Data], domain: UInt64) -> Bool {
+    static func verify(pubkeys: [Data], messages: [Data], signatures: [Data], domain: Int) -> Bool {
         return true // @todo
     }
